+++ web/src/components/process/Dashboard.vue
@@ -41,6 +41,7 @@
 
                     <most-used-forms></most-used-forms>
                     <additional-forms></additional-forms>
+                    <representation-address-forms></representation-address-forms>
                     <need-help class="mt-3"></need-help>
 
                 </b-col>
@@ -97,6 +98,7 @@ import StartEfiling from "@/components/process/AppealProcess/StartEfiling.vue";
 import NeedHelp from "@/components/utils/NeedHelp.vue";
 import MostUsedForms from "@/components/utils/MostUsedForms.vue";
 import AdditionalForms from "@/components/utils/AdditionalForms.vue";
+import RepresentationAddressForms from "@/components/utils/RepresentationAddressForms.vue";
 import { toggleStep, toggleAllSteps} from '@/components/utils/StepsPagesFunctions';
 import {GetFilingLocations} from '@/components/utils/GetFilingLocations';
 
@@ -120,7 +122,8 @@ import { form20FormsJsonDataType } from '@/types/Information/Form20';
         StartEfiling,
         NeedHelp,
         MostUsedForms,
-        AdditionalForms
+        AdditionalForms,
+        RepresentationAddressForms
     }
 })
 export default class DashboardPage extends Vue {
