+++ web/src/components/process/Form6/components/Form6StyleOfProceeding.vue
@@ -290,8 +290,8 @@ export default class Form6StyleOfProceeding extends Vue {
             form6Data.respondents = this.partiesJson.respondents;
             form6Data.formSevenNumber = this.fileNumber;
             
-            form6Data.version = this.$store.state.Application.version;            
-
+            form6Data.version = this.$store.state.Application.version;  
+            form6Data.selfRepresented = this.$store.state.Common.userSelfRepresented;
             form6Data.judgeName = Vue.filter('getFullName')(this.previousCourts[0]?.JudgeFirstName, this.previousCourts[0]?.JudgeLastName) 
             form6Data.orderDate = this.previousCourts[0]?.JudgmentDate;
             form6Data.initiatingDocumentDate = this.initiatingDocuments[0]?.DateFiled;
