+++ web/src/components/utils/MostUsedForms.vue
@@ -16,22 +16,7 @@
                 @click="startNewForm5Document"
                 target="_blank">
                 Notice of Hearing of Appeal (Form 5)                
-            </li>
-            <li class="link-button" 
-                @click="startNewForm18Document"
-                target="_blank">
-                Notice of Change of Representation/Change of Address for Service (Form 18)                
-            </li>  
-            <li class="link-button" 
-                @click="startNewForm19Document"
-                target="_blank">
-                Notice of Withdrawal of Lawyer (Form 19)                
-            </li> 
-            <li class="link-button" 
-                @click="startNewForm20Document"
-                target="_blank">
-                Notice of Objection to Withdrawal (Form 20)                
-            </li>       
+            </li>           
             <li>Certificate of Readiness</li>
             <li>Affidavit</li>
         </ol>
@@ -49,16 +34,7 @@
     const form5State = namespace("Form5");
 
     import "@/store/modules/forms/form7";
-    const form7State = namespace("Form7");
-
-    import "@/store/modules/forms/form18";
-    const form18State = namespace("Form18");
-
-    import "@/store/modules/forms/form19";
-    const form19State = namespace("Form19");
-
-    import "@/store/modules/forms/form20";
-    const form20State = namespace("Form20");
+    const form7State = namespace("Form7");   
 
     @Component
     export default class MostUsedForms extends Vue {        
@@ -70,16 +46,7 @@
 	    public UpdateCurrentNoticeOfHearingOfAppealId!: (newCurrentNoticeOfHearingOfAppealId: string) => void        
 
         @form7State.Action
-        public UpdateCurrentNoticeOfAppealId!: (newCurrentNoticeOfAppealId: string) => void
-
-        @form18State.Action
-        public UpdateCurrentNoticeOfRepChangeAddressId!: (newCurrentNoticeOfRepChangeAddressId: string) => void
-
-        @form19State.Action
-        public UpdateCurrentNoticeOfWithdrawalOfLawyerId!: (newCurrentNoticeOfWithdrawalOfLawyerId: string) => void
-
-        @form20State.Action
-        public UpdateCurrentNoticeOfObjectionToWithdrawalId!: (newCurrentNoticeOfObjectionToWithdrawalId: string) => void
+        public UpdateCurrentNoticeOfAppealId!: (newCurrentNoticeOfAppealId: string) => void      
                 
         public startNewForm2Document(){
             this.UpdateCurrentCaseId(null);
@@ -94,22 +61,7 @@
         public startNewForm7Document(){
             this.UpdateCurrentNoticeOfAppealId(null);
             this.$router.push({name: "checklist-form7" })
-        }
-
-        public startNewForm18Document(){
-            this.UpdateCurrentNoticeOfRepChangeAddressId(null);
-            this.$router.push({name: "start-form18" })
-        }
-
-        public startNewForm19Document(){
-            this.UpdateCurrentNoticeOfWithdrawalOfLawyerId(null);
-            this.$router.push({name: "start-form19" })
-        }
-
-        public startNewForm20Document(){
-            this.UpdateCurrentNoticeOfObjectionToWithdrawalId(null);
-            this.$router.push({name: "start-form20" })
-        }
+        }       
 
     }
